split Def interface from examples
